requirements.txt
# Django
Django==5.0.1

# Django REST Framework
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Django Tenants
django-tenants==3.6.1

# Database
psycopg2-binary==2.9.9

# CORS Headers
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.1

# Utilities
python-dotenv==1.0.1
